Fixed the preview pane blanking issue
The problem was that GetHtmlPreviewFile() was being called
twice and the second time an exception would be thrown
because the preview html file would be in use by another
process, thus causing null to be set on the browser control.

Another solution might be to use a reliable way to calculate
a temporary file name for the message and then simply check
if the file already exists and simply return.

The problem with using GetHashCode() for this is that
MimeMessage does not override GetHashCode(), so you just get
the object's hash code which is not necessarily unique based
on the message.