Responsive Delete button
The Delete button is now responsive if large quantities of mails are selected.

This is implemented using a StackPanel around the Buttons, giving the Delete button a MinWidth instead of a Width and setting all Margins on the buttons to 4,0,0,9 because the margin is now relative to the other (previous) button.